create useUpdateShopperContext hook
